+++ DigitDisplay/digit_display.h
@@ -15,16 +15,16 @@ public:
     ~DigitDisplay();
 
     // Sets the digit to be displayed.
-    void SetDigit(uint8_t digit);
+    void SetDigit(unsigned int digit);
     // Returns the currently displaying digit.
-    inline uint8_t GetCurrentDigit() { return current_digit; }
+    inline unsigned int GetCurrentDigit() { return current_digit; }
     // Quickly flashes the display 5 times.
     void FlashDisplay();
 
 private:
 
     // Value which we are currently displaying.
-    uint8_t current_digit;
+    unsigned int current_digit;
     // GPIO objects corresponding to the digit display.
     GPIO left_digit_gpio;
     GPIO right_digit_gpio;
